+++ src/PostgREST/Config/Database.hs
@@ -30,7 +30,7 @@ import Protolude
 
 type RoleSettings     = (HM.HashMap ByteString (HM.HashMap ByteString ByteString))
 type RoleIsolationLvl = HM.HashMap ByteString SQL.IsolationLevel
-type TimezoneNames    = Set ByteString -- cache timezone names for prefer timezone=
+type TimezoneNames    = Set Text -- cache timezone names for prefer timezone=
 
 toIsolationLevel :: (Eq a, IsString a) => a -> SQL.IsolationLevel
 toIsolationLevel a = case a of